+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - The `min/` folder has been removed. `dropzone.min.js` is now the only
   file that is minimized.
 - Remove `Dropzone.extend` and replace by the `just-extend` package.
+- There is no more `Dropzone.version`.
 
 ## 5.9.3